# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Streamr Operator management service API in Go using Gin framework.",
        "title": "Streamr Operator Service",
        "contact": {
            "name": "Joseph Kuhnsman",
            "email": "admin@ftkuhnsman.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cronjobs/create": {
            "post": {
                "description": "Adds a new cron job to the scheduler and saves it to the storage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob"
                ],
                "summary": "Create a new cron job",
                "parameters": [
                    {
                        "description": "Create Cron Job",
                        "name": "cronJob",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CronJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CronJob"
                        }
                    }
                }
            }
        },
        "/cronjobs/delete/{id}": {
            "post": {
                "description": "Deletes a cron job in the scheduler and saves the change to the storage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob"
                ],
                "summary": "Delete a cron job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cron Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CronJob"
                        }
                    }
                }
            }
        },
        "/cronjobs/disable/{id}": {
            "post": {
                "description": "Disables a cron job in the scheduler and saves it to the storage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob"
                ],
                "summary": "Disable a cron job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cron Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CronJob"
                        }
                    }
                }
            }
        },
        "/cronjobs/enable/{id}": {
            "post": {
                "description": "Enables a cron job in the scheduler and saves it to the storage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob"
                ],
                "summary": "Enable a cron job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cron Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CronJob"
                        }
                    }
                }
            }
        },
...
        "/operator/reducestaketo/{sponsorship}/{amount}": {
            "get": {
                "description": "Responds with the transaction hash.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Change Streamr Operator stake on a given sponsor.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sponsorship address",
                        "name": "sponsorship",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "amount in wei",
                        "name": "amount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/operator/stake/{sponsorship}/{amount}": {
            "get": {
                "description": "Responds with the transaction hash.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Increase Streamr Operator stake on a given sponsor.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sponsorship address",
                        "name": "sponsorship",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "amount in wei",
                        "name": "amount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/operator/stakedinto/{address}": {
            "get": {
                "description": "Responds with the Operator stake deployed in sponsorship.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get the Streamr Operator staked balance in sponsorship.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get deployed stake by sponsorship",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.StakedIntoResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "big.Int": {
            "type": "object"
        },
...
        "models.CronJob": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "endpoint": {
                    "type": "string",
                    "example": "/api/v1/sample-endpoint"
                },
                "method": {
                    "type": "string",
                    "example": "GET"
                },
                "name": {
                    "type": "string",
                    "example": "sample job"
                },
                "schedule": {
                    "type": "string",
                    "example": "0/5 * * * * *"
                }
            }
        },
...
        "models.StakedIntoResponse": {
            "type": "object",
            "properties": {
                "stakedInto": {
                    "$ref": "#/definitions/big.Int"
                }
            }
        }
    }
}